requirements.txt
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
airportsdata==20250224
annotated-types==0.7.0
anyio==4.9.0
astor==0.8.1
attrs==24.2.0
av==14.2.0
babel==2.17.0
blake3==1.0.4
blis==1.2.1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpathlib==0.21.0
cloudpickle==3.1.1
cn2an==0.5.23
colorama==0.4.6
coloredlogs==15.0.1
compressed-tensors==0.9.2
confection==0.1.5
contourpy==1.3.1
csvw==3.5.1
ctranslate2==4.4.0
cupy-cuda12x==13.4.1
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
depyf==0.18.0
dill==0.3.9
diskcache==5.6.3
distro==1.9.0
dlinfo==2.0.0
dnspython==2.7.0
docopt==0.6.2
docstring_parser==0.16
einops==0.8.1
email_validator==2.2.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
espeakng-loader==0.2.4
eval_type_backport==0.2.2
faiss-cpu==1.8.0.post1
fastapi==0.115.6
fastapi-cli==0.0.7
faster-whisper==1.1.1
fastrlock==0.8.3
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.2
fugashi==1.4.0
gguf==0.10.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-speech==2.31.1
google-cloud-texttospeech==2.25.1
google-genai==1.3.0
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
hf-xet==1.0.3
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.30.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
inflect==7.5.0
interegular==0.3.3
isodate==0.7.2
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jiter==0.7.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kokoro==0.9.2
-e git+https://github.com/remsky/Kokoro-FastAPI.git@f1fa3404940e1422010b9be19f9b4996619e5e57#egg=kokoro_fastapi
langchain==0.3.7
langchain-core==0.3.15
langchain-ollama==0.2.0
langchain-openai==0.2.5
langchain-text-splitters==0.3.2
langcodes==3.5.0
langsmith==0.1.139
language-tags==1.2.0
language_data==1.3.0
lark==1.2.2
livekit==1.0.6
livekit-agents==1.0.13
livekit-api==1.0.2
livekit-plugins-cartesia==0.4.11
livekit-plugins-deepgram==1.0.13
livekit-plugins-google==0.11.2
livekit-plugins-noise-cancellation==0.2.1
livekit-plugins-openai==1.0.13
livekit-plugins-silero==1.0.13
livekit-plugins-turn-detector==1.0.13
livekit-protocol==1.0.2
llguidance==0.7.14
llvmlite==0.44.0
lm-format-enforcer==0.10.11
loguru==0.7.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
memoripy==0.1.2
misaki==0.9.3
mistral_common==1.5.4
mojimoji==0.0.13
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
munch==4.0.0
murmurhash==1.0.12
mutagen==1.47.0
nanobind==2.6.1
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
ollama==0.3.3
onnxruntime==1.21.0
openai==1.74.1
opencv-python-headless==4.11.0.86
ordered-set==4.1.0
orjson==3.10.11
outlines==0.1.11
outlines_core==0.1.26
packaging==24.1
partial-json-parser==0.2.1.1.post5
phonemizer-fork==3.3.2
pillow==11.1.0
preshed==3.0.9
proces==0.1.7
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.21.1
propcache==0.2.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.10.4
pydantic-settings==2.7.0
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
PyJWT==2.10.1
pyopenjtalk-plus==0.3.4.post11
pyparsing==3.2.3
pypinyin==0.54.0
pypinyin-dict==0.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
PyYAML==6.0.2
pyzmq==26.4.0
ray==2.43.0
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==1.5.0
rich==14.0.0
rich-toolkit==0.14.1
rpds-py==0.24.0
rsa==4.9
safetensors==0.5.3
scikit-learn==1.5.2
scipy==1.14.1
segments==2.3.0
sentencepiece==0.2.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sounddevice==0.5.1
soundfile==0.13.0
spacy==3.8.5
spacy-curated-transformers==0.3.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.27
srsly==2.5.1
starlette==0.41.3
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.1
tenacity==9.0.0
text2num==2.5.1
thinc==8.3.4
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.1
torch==2.6.0+cu124
torchaudio==2.6.0
torchvision==0.21.0
tqdm==4.67.1
transformers==4.51.2
triton==3.2.0
typeguard==4.4.2
typer==0.15.2
types-protobuf==4.25.0.20240417
typing-inspection==0.4.0
typing_extensions==4.13.2
unidic-lite==1.0.8
uritemplate==4.1.1
urllib3==2.2.3
uv==0.6.12
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.8.3
wasabi==1.1.3
watchfiles==1.0.4
weasel==0.4.1
websockets==14.2
wrapt==1.17.2
xformers==0.0.29.post2
xgrammar==0.1.17
yarl==1.17.1
zipp==3.21.0
